add "area" visualization
the popup was intended to hold other visualizations, but those menu
options have been moved into the unified view.html page. so the popup
is gone, but a background.js is needed to react to clicks on the
browser action icon.

the new visualization is written in scala-js, and the trampoli3n.js
functionality was moved into the src.scala file. the chromi3wm-opt
files are what scala-js compiles src.scala into and are included in
the repository to avoid the added sbt dependency. but a Makefile and
all the necessary sbt project files rebuild the opt files as needed.